refactor: make changes in calendar
