+++ src/views/dashboard/index.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 import DashboardLayout from "../../layouts/dashboard";
 import SidebarLayout from "../../layouts/sidebar";
-import CalendarContainer from "../../components/containers/CalendarContainer";
-
+import CalendarContainer from "../../containers/calendarContainer";
+import TaskContainer from "../../containers/taskContainer";
 const DashboardPage = () => {
 	useEffect(() => {
 		document.title = "Qlegal - Dashboard";
@@ -14,6 +14,8 @@ const DashboardPage = () => {
 				<div className="container flex flex-col px-4 lg:flex-row lg:gap-8">
 					<div className="w-full md:w-1/2">
 						<CalendarContainer />
+
+						<TaskContainer />
 					</div>
 				</div>
 			</SidebarLayout>
